Делит main.js на модульные файлы
